+++ src/commands/new_stacked_view_view_model/index.ts
@@ -1,2 +1,1 @@
 export * from "./new_stacked_view_view_model";
-export * from "./new_stacked_skeleton";