+++ src/templates/index.ts
@@ -1,2 +1,1 @@
-export * from "./stacked_view.template";
-export * from "./stacked_view_model.template";
+export * from "./stacked_view_view_model";