+++ src/utils/index.ts
@@ -1,5 +1,3 @@
 export * from "./create_directory";
-export * from "./create_stacked_templates";
 export * from "./prompt_for";
-export * from "./generate_stacked_code";
-export * from "./helpers";
+export * from "./utils";