+++ CHANGELOG.md
@@ -6,6 +6,11 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 
 ## [Unreleased]
 
+## Added
+
+- Changed name of `Stacked: New Stacked` command to `Stacked: New Stacked View ViewModel` command
+- `Stacked: New Stacked Skeleton` command
+
 ### Fixed
 
 - Code repetition